Bug fixes, esp. related to re-running get-latest after interruption
Before these changes, get-latest would always go until the most recent
item downloaded from the given account. This potentially skips items
if get-latest was interrupted, then run again later, because get-latest
will stop once it finds the most recent item downloaded, usually one
of the first things downloaded.

So, this adds a cursor/marker to the DB for the account so that we know
which item ID was most recent as part of the last successful run; that
way, interrupted runs will not move the cursor, and thus no items will
be lost simply because get-latest was too naive to know that it should
keep scanning until an older timestamp.